refactor (INTERPOLATE) #51 move NumberTests to common module
